Move Enumerable to rely less on shared examples
