+++ core/enumerable/collect_concat_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/collect_concat'
 
 describe "Enumerable#collect_concat" do
-  it_behaves_like :enumerable_collect_concat, :collect_concat
+  it "is an alias of Enumerable#flat_map" do
+    Enumerable.instance_method(:collect_concat).should == Enumerable.instance_method(:flat_map)
+  end
 end
