+++ core/enumerable/collect_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/collect'
 
 describe "Enumerable#collect" do
-  it_behaves_like :enumerable_collect, :collect
+  it "is an alias of Enumerable#map" do
+    Enumerable.instance_method(:collect).should == Enumerable.instance_method(:map)
+  end
 end
