+++ core/enumerable/detect_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/find'
 
 describe "Enumerable#detect" do
-  it_behaves_like :enumerable_find, :detect
+  it "is an alias of Enumerable#find" do
+    Enumerable.instance_method(:detect).should == Enumerable.instance_method(:find)
+  end
 end
