+++ core/enumerable/entries_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/entries'
 
 describe "Enumerable#entries" do
-  it_behaves_like :enumerable_entries, :entries
+  it "is an alias of Enumerable#to_a" do
+    Enumerable.instance_method(:entries).should == Enumerable.instance_method(:to_a)
+  end
 end
