+++ core/enumerable/filter_spec.rb
@@ -1,7 +1,7 @@
 require_relative '../../spec_helper'
-require_relative 'fixtures/classes'
-require_relative 'shared/find_all'
 
 describe "Enumerable#filter" do
-  it_behaves_like :enumerable_find_all, :filter
+  it "is an alias of Enumerable#select" do
+    Enumerable.instance_method(:filter).should == Enumerable.instance_method(:select)
+  end
 end
