+++ core/enumerable/map_spec.rb
@@ -1,7 +1,109 @@
 require_relative '../../spec_helper'
 require_relative 'fixtures/classes'
-require_relative 'shared/collect'
+require_relative 'shared/enumerable_enumeratorized'
 
 describe "Enumerable#map" do
-  it_behaves_like :enumerable_collect, :map
+  before :each do
+    ScratchPad.record []
+  end
+
+  it "returns a new array with the results of passing each element to block" do
+    entries = [0, 1, 3, 4, 5, 6]
+    numerous = EnumerableSpecs::Numerous.new(*entries)
+    numerous.map { |i| i % 2 }.should == [0, 1, 1, 0, 1, 0]
+    numerous.map { |i| i }.should == entries
+  end
+
+  it "passes through the values yielded by #each_with_index" do
+    [:a, :b].each_with_index.map { |x, i| ScratchPad << [x, i]; nil }
+    ScratchPad.recorded.should == [[:a, 0], [:b, 1]]
+  end
+
+  it "gathers initial args as elements when each yields multiple" do
+    multi = EnumerableSpecs::YieldsMulti.new
+    multi.map {|e| e}.should == [1,3,6]
+  end
+
+  it "only yields increasing values for a Range" do
+    (1..0).map { |x| x }.should == []
+    (1..1).map { |x| x }.should == [1]
+    (1..2).map { |x| x }.should == [1, 2]
+  end
+
+  it "returns an enumerator when no block given" do
+    enum = EnumerableSpecs::Numerous.new.map
+    enum.should.instance_of?(Enumerator)
+    enum.each { |i| -i }.should == [-2, -5, -3, -6, -1, -4]
+  end
+
+  it "reports the same arity as the given block" do
+    entries = [0, 1, 3, 4, 5, 6]
+    numerous = EnumerableSpecs::Numerous.new(*entries)
+
+    def numerous.each(&block)
+      ScratchPad << block.arity
+      super
+    end
+
+    numerous.map { |a, b| a % 2 }.should == [0, 1, 1, 0, 1, 0]
+    ScratchPad.recorded.should == [2]
+    ScratchPad.clear
+    ScratchPad.record []
+    numerous.map { |i| i }.should == entries
+    ScratchPad.recorded.should == [1]
+  end
+
+  it "yields an Array of 2 elements for a Hash when block arity is 1" do
+    c = Class.new do
+      def register(a)
+        ScratchPad << a
+      end
+    end
+    m = c.new.method(:register)
+
+    ScratchPad.record []
+    { 1 => 'a', 2 => 'b' }.map(&m)
+    ScratchPad.recorded.should == [[1, 'a'], [2, 'b']]
+  end
+
+  it "yields 2 arguments for a Hash when block arity is 2" do
+    c = Class.new do
+      def register(a, b)
+        ScratchPad << [a, b]
+      end
+    end
+    m = c.new.method(:register)
+
+    ScratchPad.record []
+    { 1 => 'a', 2 => 'b' }.map(&m)
+    ScratchPad.recorded.should == [[1, 'a'], [2, 'b']]
+  end
+
+  it "raises an error for a Hash when an arity enforcing block of arity >2 is passed in" do
+    c = Class.new do
+      def register(a, b, c)
+      end
+    end
+    m = c.new.method(:register)
+
+    -> do
+      { 1 => 'a', 2 => 'b' }.map(&m)
+    end.should.raise(ArgumentError)
+  end
+
+  it "calls the each method on sub-classes" do
+    c = Class.new(Hash) do
+      def each
+        ScratchPad << 'in each'
+        super
+      end
+    end
+    h = c.new
+    h[1] = 'a'
+    ScratchPad.record []
+    h.map { |k,v| v }
+    ScratchPad.recorded.should == ['in each']
+  end
+
+  it_behaves_like :enumerable_enumeratorized_with_origin_size, :map
 end
